requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.1
PyMySQL==1.1.0
python-socketio==5.10.0

